Refactored api routes to seperate files and removed deprecated files
